rures modified
