+++ src/App.jsx
@@ -8,8 +8,9 @@ import Footer from './components/Footer/Footer';
 import Home from './pages/Home';
 import Chatbot from './pages/Chatbot';
 import Dropdown from './components/dropdown/Dropdown';
-import Profile from './components/Profile/Profile';
+import User from './components/User/User';
 import axios from 'axios';
+import SavedLocations from './pages/SavedLocations';
 
 function App() {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
@@ -40,7 +41,7 @@ function App() {
 
   const handleSignUpSuccess = (user) => {
     setIsAuthenticated(true);
-    setUsername(user.username); // Asegúrate de que `user.username` esté correctamente
+    setUsername(user.username); 
     navigate('/'); 
   };
 
@@ -60,11 +61,10 @@ function App() {
         <Route path="/login" element={<Login onLoginSuccess={handleLoginSuccess} />} />
         <Route path="/signup" element={<SignUp onSignUpSuccess={handleSignUpSuccess} />} />
         <Route path="/chatbot" element={<Chatbot />} />
-        
-        {/* Ruta protegida para el perfil */}
-        <Route path="/profile" element={isAuthenticated ? <Profile /> : <Login onLoginSuccess={handleLoginSuccess} />} />
-
+        <Route path="/user" element={<User /> } />
+        <Route path="/savedlocations" element={<SavedLocations />} />
         <Route path="/dropdown" element={<Dropdown />} />
+        
       </Routes>
 
       <Footer />
