+++ src/components/nav/nav.css
@@ -1,6 +1,6 @@
 .navbar {
   display: flex;
-  align-items: center; /* Asegura que todos los elementos se alineen verticalmente */
+  align-items: center; 
   background-color: #ffffff;
   color: #ffffff;
   position: fixed;
@@ -13,18 +13,18 @@
 
 .navbar-logo {
   display: flex;
-  align-items: center; /* Mantiene la imagen y el texto alineados verticalmente */
+  align-items: center;
 }
 
 .navbar-logo img {
   height: 5rem;
-  display: block; /* Asegura que la imagen no tenga margen extra por defecto */
+  display: block; 
 }
 
 .navbar-links {
   display: flex;
   gap: 1.5rem;
-  align-items: center; /* Asegura que los enlaces estén alineados verticalmente */
+  align-items: center; 
 }
 
 .navbar-links a.button:hover {
