+++ src/config/urls.jsx
@@ -1,3 +1,6 @@
 export const BASE_URL = 'http://127.0.0.1:8000/api/v1/';
 export const USER_REGISTER = `${BASE_URL}register/`;
 export const USER_LOGIN = `${BASE_URL}login/`;
+
+export const USER_DETAIL = `${BASE_URL}user/`;
+export const PREFERENCES = `${BASE_URL}preferences/`;
