+++ src/index.css
@@ -3,7 +3,6 @@
   body {
     margin: 0;
     padding: 0;
-    display: flex;
     justify-content: center;
     align-items: center;
     height: 100vh; /* Asegura que el body ocupe toda la altura */
