+++ src/pages/Home.jsx
@@ -4,10 +4,9 @@ import Banner from '../components/Banner/Banner';
 import { useNavigate } from 'react-router-dom';
 
 function Home() {
-  const navigate = useNavigate();  // Define useNavigate para redirigir
-
+  const navigate = useNavigate();  
   const handleChatbotClick = () => {
-    navigate('/chatbot');  // Redirige a la ruta /chatbot
+    navigate('/chatbot');  
   };
 
   const handleSearchPlanningClick = () => {
@@ -29,7 +28,7 @@ function Home() {
 
       <div className="main-buttons">
         <button className="main-button" onClick={handleSearchPlanningClick}>SEARCH PLANNING</button>
-        {/* Cambia el evento onClick del botón CHATBOT */}
+        
         <button className="main-button" onClick={handleChatbotClick}>CHATBOT</button>
       </div>
     </main>
